refactor: P2-8 PakGenericMetadataのrow builderを分離 + lintルール修正
